requirements.txt
torch
torchvision
fastapi
uvicorn
tqdm
opencv-python
Pillow